HTTP Standard eingeführt, Nutzerlöschung überarbeitet, TotalListings teils implementiert

See merge request swe1_ws_2324/projects/project_14!9

+++ spring/src/main/java/edu/hm/cs/buecherkreisel/spring/UserController.java
@@ -2,6 +2,7 @@ package edu.hm.cs.buecherkreisel.spring;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -13,8 +14,18 @@ public class UserController {
     @Autowired
     private UserRepository repository;
 
+    @Autowired
+    ListingRepository listRepo;
+
+    /**
+     * Retrieves user information based on the provided ID and token.
+     *
+     * @param id   The ID of the user to retrieve.
+     * @param body Map containing the user token for authentication.
+     * @return ResponseEntity containing user information if successful, or an error response.
+     */
     @GetMapping("/users/{id}")
-    Map<String, String> getUser(@PathVariable long id, @RequestBody Map<String, String> body) {
+    ResponseEntity<Map<String, String>> getUser(@PathVariable long id, @RequestBody Map<String, String> body) {
         Optional<User> userCheck = repository.findById(id);
         if(userCheck.isPresent()) {
             User user = userCheck.get();
@@ -24,15 +35,22 @@ public class UserController {
                 answer.put("username", user.getUsername());
                 answer.put("profile_picture", user.getProfilePicture());
                 answer.put("liked_listings", user.getLikedListings().toString());
-                return answer;
+                answer.put("total_listings", Integer.toString(user.getTotalListings()));
+                return new ResponseEntity<>(answer, HttpStatus.OK);
             } else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
         } else {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
     }
 
+    /**
+     * Retrieves public user information based on the provided ID.
+     *
+     * @param id The ID of the user to retrieve public information for.
+     * @return ResponseEntity containing public user information if successful, or an error response.
+     */
     @GetMapping("/users/{id}/public")
-    Map<String, String> getUserPublic(@PathVariable long id) {
+    ResponseEntity<Map<String, String>> getUserPublic(@PathVariable long id) {
         Optional<User> userCheck = repository.findById(id);
         if(userCheck.isPresent()) {
             User user = userCheck.get();
@@ -40,12 +58,19 @@ public class UserController {
             userElements.put("id", String.valueOf(user.getId()));
             userElements.put("username",user.getUsername());
             userElements.put("profile_picture",user.getProfilePicture());
-            return userElements;
+            userElements.put("total_listings", Integer.toString(user.getTotalListings()));
+            return new ResponseEntity<>(userElements, HttpStatus.OK);
         } else throw new ResponseStatusException(HttpStatus.NOT_FOUND);
     }
 
+    /**
+     * Creates a new user with the provided username and password.
+     *
+     * @param body Map containing the username and password for the new user.
+     * @return ResponseEntity containing the ID and token of the created user if successful, or an error response.
+     */
     @PostMapping("/users/create")
-    Map<String, String> createUser(@RequestBody Map<String, String> body) {
+    ResponseEntity<Map<String, String>> createUser(@RequestBody Map<String, String> body) {
         String username = body.get("username");
         if(repository.findAll().stream()
                 .anyMatch(user -> Objects.equals(user.getUsername(), username))) {
@@ -56,11 +81,17 @@ public class UserController {
         Map<String, String> answer = new HashMap<>();
         answer.put("id", String.valueOf(user.getId()));
         answer.put("token", user.getToken());
-        return answer;
+        return new ResponseEntity<>(answer, HttpStatus.CREATED);
     }
 
+    /**
+     * Authenticates a user based on the provided username and password.
+     *
+     * @param body Map containing the username and password for user authentication.
+     * @return ResponseEntity containing the ID and token of the authenticated user if successful, or an error response.
+     */
     @PostMapping("/login")
-    Map<String, String> login(@RequestBody Map<String, String> body) {
+    ResponseEntity<Map<String, String>> login(@RequestBody Map<String, String> body) {
         String username = body.get("username");
         String password = body.get("password");
         for(User user : repository.findAll()) {
@@ -69,15 +100,22 @@ public class UserController {
                     Map<String, String> answer = new HashMap<>();
                     answer.put("id", String.valueOf(user.getId()));
                     answer.put("token", user.getToken());
-                    return answer;
+                    return new ResponseEntity<>(answer, HttpStatus.OK);
                 } else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
             }
         }
         throw new ResponseStatusException(HttpStatus.NOT_FOUND);
     }
 
-    @PutMapping("/users/edit_likes/{id}")
-    void editLikes(@PathVariable long id, @RequestBody Map<String, String> body) {
+    /**
+     * Updates the liked listings of a user based on the provided token.
+     *
+     * @param id   The ID of the user to update.
+     * @param body Map containing the user token and updated liked listings.
+     * @return ResponseEntity indicating success or an error response.
+     */
+    @PatchMapping("/users/edit_likes/{id}")
+    ResponseEntity<?> editLikes(@PathVariable long id, @RequestBody Map<String, String> body) {
         Optional<User> userCheck = repository.findById(id);
         if(userCheck.isPresent()) {
             User user = userCheck.get();
@@ -91,12 +129,20 @@ public class UserController {
                 }
                 user.setLikedListings(listOfLikes);
                 repository.save(user);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
             } else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
         } else throw new ResponseStatusException(HttpStatus.NOT_FOUND);
     }
 
-    @PutMapping("/users/edit_auth/{id}")
-    Map<String, String> editAuth(@PathVariable long id, @RequestBody Map<String, String> body) {
+    /**
+     * Updates the authentication details of a user based on the provided token.
+     *
+     * @param id   The ID of the user to update.
+     * @param body Map containing the user token, old password, and new password.
+     * @return ResponseEntity containing the updated token if successful, or an error response.
+     */
+    @PatchMapping("/users/edit_auth/{id}")
+    ResponseEntity<Map<String, String>> editAuth(@PathVariable long id, @RequestBody Map<String, String> body) {
         Optional<User> userCheck = repository.findById(id);
         if(userCheck.isPresent()) {
             User user = userCheck.get();
@@ -108,13 +154,20 @@ public class UserController {
                 repository.save(user);
                 Map<String,String> response = new HashMap<>();
                 response.put("token", user.getToken());
-                return response;
+                return new ResponseEntity<>(response, HttpStatus.OK);
             } else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
         } else throw new ResponseStatusException(HttpStatus.NOT_FOUND);
     }
 
-    @PutMapping("/users/edit_alias/{id}")
-    void editAlias(@PathVariable long id, @RequestBody Map<String, String> body) {
+    /**
+     * Updates the username of a user based on the provided token.
+     *
+     * @param id   The ID of the user to update.
+     * @param body Map containing the user token and new username.
+     * @return ResponseEntity indicating success or an error response.
+     */
+    @PatchMapping("/users/edit_alias/{id}")
+    ResponseEntity<?> editAlias(@PathVariable long id, @RequestBody Map<String, String> body) {
         Optional<User> userCheck = repository.findById(id);
         if(userCheck.isPresent()) {
             User user = userCheck.get();
@@ -122,12 +175,20 @@ public class UserController {
             if(user.getToken().equals(token)){
                 user.setUsername(body.get("new_alias"));
                 repository.save(user);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
             } else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
         } else throw new ResponseStatusException(HttpStatus.NOT_FOUND);
     }
 
-    @PutMapping("/users/edit_pic/{id}")
-    void editPicture(@PathVariable long id, @RequestBody Map<String, String> body) {
+    /**
+     * Updates the profile picture URL of a user based on the provided token.
+     *
+     * @param id   The ID of the user to update.
+     * @param body Map containing the user token and new profile picture URL.
+     * @return ResponseEntity indicating success or an error response.
+     */
+    @PatchMapping("/users/edit_pic/{id}")
+    ResponseEntity<?> editPicture(@PathVariable long id, @RequestBody Map<String, String> body) {
         Optional<User> userCheck = repository.findById(id);
         if(userCheck.isPresent()) {
             User user = userCheck.get();
@@ -135,12 +196,20 @@ public class UserController {
             if(user.getToken().equals(token)) {
                 user.setProfilePicture(body.get("new_picture"));
                 repository.save(user);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
             } else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
         } else throw new ResponseStatusException(HttpStatus.NOT_FOUND);
     }
 
+    /**
+     * Deletes a user and associated listings based on the provided token and password.
+     *
+     * @param id   The ID of the user to delete.
+     * @param body Map containing the user token and password for deletion authentication.
+     * @return ResponseEntity indicating success or an error response.
+     */
     @DeleteMapping("/users/delete/{id}")
-    void deleteUser(@PathVariable long id, @RequestBody Map<String, String> body) {
+    ResponseEntity<?> deleteUser(@PathVariable long id, @RequestBody Map<String, String> body) {
         Optional<User> userCheck = repository.findById(id);
         if(userCheck.isPresent()) {
             User user = userCheck.get();
@@ -148,6 +217,10 @@ public class UserController {
             String password = body.get("password");
             if(user.getToken().equals(token) && user.getPassword().equals(password)) {
                 repository.deleteById(id);
+                List<Listing> userListings = listRepo.findAll().stream()
+                        .filter(listing -> listing.getUserID().equals(id)).toList();
+                userListings.forEach(listing -> listRepo.deleteById(listing.getId()));
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
             } else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
         } else throw new ResponseStatusException(HttpStatus.NOT_FOUND);
     }
